requirements.txt
altgraph==0.17.3
auto-py-to-exe==2.36.0
bottle==0.12.25
bottle-websocket==0.2.9
cffi==1.15.1
Eel==0.16.0
ffmpeg-python==0.2.0
future==0.18.3
gevent==22.10.2
gevent-websocket==0.10.1
greenlet==2.0.2
pefile==2023.2.7
pycparser==2.21
pyinstaller==5.12.0
pyinstaller-hooks-contrib==2023.3
pyparsing==3.1.0
pytube==15.0.0
pywin32-ctypes==0.2.1
whichcraft==0.6.1
zope.event==4.6
zope.interface==6.0
